Moved test files and improved image testing
- Moved each testfile to a folder corresponding to the node they test
- Moved image files to images folder
- Improved tests so they can run on multiple images
