+++ catkin_ws/src/acr_core/test/test_cameraToAngle/images/left/0.png
