+++ catkin_ws/src/acr_core/test/test_cameraToAngle/images/none/0.png
