+++ catkin_ws/src/acr_core/test/test_cameraToAngle/images/right/0.png
